Moved bot-related code from Main to Bot
